doc: added commens
