+++ http/auth.go
@@ -11,26 +11,30 @@ import (
 	"github.com/google/uuid"
 )
 
-// The claims to generate a Hagall User JWT token.
+// HagallUserClaim is the claims to generate a Hagall User JWT token.
 type HagallUserClaim struct {
 	jwt.RegisteredClaims
 
 	AppKey string `json:"app_key"`
 }
 
+// MakeJWTSecret creates a random secret string
 func MakeJWTSecret() string {
 	return base64.RawURLEncoding.EncodeToString([]byte(uuid.NewString()))
 }
 
+// MakeAuthorizationHeader creates a Bearer authorization header with toke
 func MakeAuthorizationHeader(token string) string {
 	return "Bearer " + token
 }
 
+// SignIdentity signs endpoint with secret
 func SignIdentity(endpoint, secret string) (string, error) {
 	token := jwt.NewWithClaims(jwt.SigningMethodHS256, jwt.MapClaims{"endpoint": endpoint})
 	return token.SignedString([]byte(secret))
 }
 
+// VerifyHagallUserAccessToken verifies signed token with the secret
 func VerifyHagallUserAccessToken(token, secret string) error {
 	var claims HagallUserClaim
 
@@ -55,8 +59,7 @@ func VerifyHagallUserAccessToken(token, secret string) error {
 	return err
 }
 
-// GenerateHagallUserAccessToken generate a Hagall user access token with the
-// given secret.
+// GenerateHagallUserAccessToken generate a Hagall user access token with the given secret.
 func GenerateHagallUserAccessToken(appKey, secret string, ttl time.Duration) (string, error) {
 	now := time.Now()
 
@@ -74,6 +77,7 @@ func GenerateHagallUserAccessToken(appKey, secret string, ttl time.Duration) (st
 	return token.SignedString([]byte(secret))
 }
 
+// GetAppKeyFromHTTPRequest extracts app_key from the http request authorization header
 func GetAppKeyFromHTTPRequest(r *http.Request) string {
 	switch auth := r.Header.Get("Authorization"); {
 	case strings.HasPrefix(auth, "Basic"):
